plugins/requirements.txt
beautifulsoup4 # bs4
lxml # for parsing html
Pillow>=8.1.1 # PIL
PyPDF2
